# api/docs/swagger.yaml
basePath: /
definitions:
  userservice.LoginUserRequest:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  userservice.RefReshTokenRes:
    properties:
      message:
        type: string
    type: object
  userservice.RegisterUserRequest:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  userservice.RegisterUserResponse:
    properties:
      user:
        $ref: '#/definitions/userservice.User'
    type: object
  userservice.User:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: string
      password:
        type: string
      role:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
host: localhost:50052
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server for a user service.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: User Service API
  version: "1.0"
paths:
  /login:
    post:
      consumes:
      - application/json
      description: Login user with username and password
      parameters:
      - description: User data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/userservice.LoginUserRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Login user
      tags:
      - User
  /refresh-token:
    post:
      consumes:
      - application/json
      description: This endpoint refreshes the signing key and returns a confirmation
        message.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/userservice.RefReshTokenRes'
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Refreshe Token
      tags:
      - authentication
  /register:
    post:
      consumes:
      - application/json
      description: Register a new user with username and password
      parameters:
      - description: User data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/userservice.RegisterUserRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/userservice.RegisterUserResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Register a new user
      tags:
      - User
swagger: "2.0"